4runr-agents/scraper/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
colorama>=0.4.4
playwright==1.40.0